# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PowerHub — multi-tenant hotel power/relay automation control system
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rooms:
    get:
      operationId: listRooms
      tags: [rooms]
      parameters:
        - $ref: "#/components/parameters/PropertyIdQuery"
      responses:
        "200":
          description: Rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"
    post:
      operationId: createRoom
      tags: [rooms]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "400":
          $ref: "#/components/responses/BadRequest"
  /rooms/{id}:
    patch:
      operationId: updateRoom
      tags: [rooms]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteRoom
      tags: [rooms]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppUser"
    post:
      operationId: createUser
      tags: [users]
      summary: Register an application user (by email) with role and property allocation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppUserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"
        "400":
          $ref: "#/components/responses/BadRequest"
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppUserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteUser
      tags: [users]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
    PropertyIdQuery:
      name: propertyId
      in: query
      required: true
      schema:
        type: integer
  responses:
...
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Room:
      type: object
      required: [id, propertyId, roomNo, active]
      properties:
        id: { type: integer }
        propertyId: { type: integer }
        roomNo: { type: string }
        blockId: { type: ["integer", "null"] }
        floorId: { type: ["integer", "null"] }
        roomTypeId: { type: ["integer", "null"] }
        blockName: { type: ["string", "null"] }
        floorName: { type: ["string", "null"] }
        roomTypeName: { type: ["string", "null"] }
        active: { type: boolean }
    RoomInput:
      type: object
      required: [propertyId, roomNo]
      properties:
        propertyId: { type: integer }
        roomNo: { type: string, minLength: 1 }
        blockId: { type: integer }
        floorId: { type: integer }
        roomTypeId: { type: integer }
        active: { type: boolean }
    RoomUpdate:
      type: object
      properties:
        roomNo: { type: string, minLength: 1 }
        blockId: { type: ["integer", "null"] }
        floorId: { type: ["integer", "null"] }
        roomTypeId: { type: ["integer", "null"] }
        active: { type: boolean }
...
    AppUser:
      type: object
      required: [id, email, name, isSuperAdmin, active, propertyIds]
      properties:
        id: { type: integer }
        email: { type: string }
        name: { type: string }
        isSuperAdmin: { type: boolean }
        active: { type: boolean }
        roleId: { type: ["integer", "null"] }
        roleName: { type: ["string", "null"] }
        propertyIds:
          type: array
          items: { type: integer }
        createdAt: { type: string, format: date-time }
    AppUserInput:
      type: object
      required: [email, name]
      properties:
        email: { type: string, minLength: 1 }
        name: { type: string, minLength: 1 }
        roleId: { type: integer }
        isSuperAdmin: { type: boolean }
        active: { type: boolean }
        propertyIds:
          type: array
          items: { type: integer }
    AppUserUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        roleId: { type: ["integer", "null"] }
        isSuperAdmin: { type: boolean }
        active: { type: boolean }
        propertyIds:
          type: array
          items: { type: integer }
...